refactor(service): 纯移动拆分 antigravity_gateway_service.go（4664→639行）
